Serious Refactoring, code not quite operational
Creating a more MVC- oriented code design, so that everything is
actually organized.  Data comes exclusively from ConnectedService (in
obdService), where it is then messaged to and interpreted in
obdService,, which then sends only the data to be written to the screen
to the UI thread (MainActivity).  Overall much cleaner.